fixed padding after "use strict"; in all js files.

+++ scripts/ipc-handlers/projectGetters.js
@@ -4,10 +4,17 @@ require("../extend/console.js"); // Adds more verbose logging to the console and
 const { ipcMain } = require("electron");
 const fs = require("node:fs");
 const path = require("node:path")
-const { saveFilesPath } = require("../../main.js");
+const { saveFilesFolder } = require("../../main.js");
 
 const projectGet = (id) => {
-    const filePath = path.join(saveFilesPath, id + ".json")
+    console.info(`Getting project: ${id}`);
+
+    if (!id) {
+        console.error("No ID provided for the project get");
+        return null;
+    }
+
+    const filePath = path.join(saveFilesFolder, id + ".json")
 
     if (!fs.existsSync(filePath)) {
         return null;
@@ -18,28 +25,41 @@ const projectGet = (id) => {
     if (data) {
         return data;
     } else {
-        console.error("Error reading project file: " + filePath);
+        console.error(`Error reading project file: ${filePath}`);
         return null;
     }
 };
 
 const projectSave = (projectData) => {
-    console.info("Saving project: " + projectData.date + " - " + projectData.id);
-    const id = projectData.id;
-    const filePath = path.join(saveFilesPath, projectData.id + ".json");
+    console.info(`Saving project: ${projectData.date} - ${projectData.id}`);
 
+    if (!projectData) {
+        console.error("No project data provided for the project save");
+        return null;
+    }
+
+    const id = projectData.id;
     if (!id) {
         console.error("No ID provided for the project save");
         return null;
     }
 
+    const filePath = path.join(saveFilesFolder, id + ".json");
+    
     fs.writeFileSync(filePath, JSON.stringify(projectData), { flag: "w" });
 
     return "Project saved";
 };
 
 const projectDelete = (id) => {
-    const filePath = path.join(saveFilesPath, id + ".json")
+    console.info(`Deleting project: ${id}`);
+
+    if (!id) {
+        console.error("No ID provided for the project delete");
+        return null;
+    }
+
+    const filePath = path.join(saveFilesFolder, id + ".json")
 
     if (fs.existsSync(filePath)) {
         fs.rmSync(filePath);
@@ -50,10 +70,12 @@ const projectDelete = (id) => {
 };
 
 const projectGetAll = () => {
-    const projectFiles = fs.readdirSync(saveFilesPath);
+    console.info("Getting all projects");
+
+    const projectFiles = fs.readdirSync(saveFilesFolder);
 
     return projectFiles.map((file) => {
-        return JSON.parse(fs.readFileSync(path.join(saveFilesPath, file), { encoding: "utf-8" }));
+        return JSON.parse(fs.readFileSync(path.join(saveFilesFolder, file), { encoding: "utf-8" }));
     });
 };
 
@@ -61,8 +83,8 @@ const ipcHandlers = () => {
     ipcMain.handle("project-get", (_, id) => {
         return projectGet(id);
     });
-    ipcMain.handle("project-save", (_, projectJSON) => {
-        return projectSave(projectJSON);
+    ipcMain.handle("project-save", (_, projectData) => {
+        return projectSave(projectData);
     });
     ipcMain.handle("project-delete", (_, id) => {
         return projectDelete(id);
